service2/app/requirements.txt
fastapi==0.68.0
dapr-ext-fastapi==1.10.0